refactor: remove sqllog and move it to limited
