+++ pkg/limited/dbsize.go
@@ -1,7 +1,20 @@
 package limited
 
-import "context"
+import (
+	"context"
+	"fmt"
 
-func (l *LimitedServer) dbSize(ctx context.Context) (int64, error) {
-	return l.backend.DbSize(ctx)
+	"go.opentelemetry.io/otel/attribute"
+)
+
+func (l *LimitedServer) DbSize(ctx context.Context) (int64, error) {
+	var err error
+	ctx, span := otelTracer.Start(ctx, fmt.Sprintf("%s.DbSize", otelName))
+	defer func() {
+		span.RecordError(err)
+		span.End()
+	}()
+	size, err := l.driver.GetSize(ctx)
+	span.SetAttributes(attribute.Int64("size", size))
+	return size, err
 }
