+++ pkg/limited/get.go
@@ -23,7 +23,7 @@ func (l *LimitedServer) get(ctx context.Context, r *etcdserverpb.RangeRequest) (
 		attribute.Int64("revision", r.Revision),
 	)
 
-	rev, kv, err := l.backend.List(ctx, r.Key, nil, 1, r.Revision)
+	rev, kv, err := l.InternalList(ctx, r.Key, nil, 1, r.Revision)
 	if err != nil {
 		return nil, err
 	}
