+++ pkg/limited/otel.go
@@ -12,13 +12,15 @@ import (
 const otelName = "limited-server"
 
 var (
-	otelTracer trace.Tracer
-	otelMeter  metric.Meter
-	createCnt  metric.Int64Counter
-	deleteCnt  metric.Int64Counter
-	getCnt     metric.Int64Counter
-	listCnt    metric.Int64Counter
-	updateCnt  metric.Int64Counter
+	otelTracer      trace.Tracer
+	otelMeter       metric.Meter
+	createCnt       metric.Int64Counter
+	deleteCnt       metric.Int64Counter
+	getCnt          metric.Int64Counter
+	listCnt         metric.Int64Counter
+	updateCnt       metric.Int64Counter
+	compactCnt      metric.Int64Counter
+	watcherGroupCnt metric.Int64Counter
 )
 
 func init() {
@@ -45,5 +47,13 @@ func init() {
 	updateCnt, err = otelMeter.Int64Counter(fmt.Sprintf("%s.update", otelName), metric.WithDescription("Number of update requests"))
 	if err != nil {
 		logrus.WithError(err).Warning("Otel failed to create update counter")
+	}
+	compactCnt, err = otelMeter.Int64Counter(fmt.Sprintf("%s.compact", otelName), metric.WithDescription("Number of compact requests"))
+	if err != nil {
+		logrus.WithError(err).Warning("Otel failed to create create counter")
+	}
+	watcherGroupCnt, err = otelMeter.Int64Counter(fmt.Sprintf("%s.watcherGroup", otelName), metric.WithDescription("Number of watcherGroup requests"))
+	if err != nil {
+		logrus.WithError(err).Warning("Otel failed to create create counter")
 	}
 }
