--- pkg/sqllog/sqllog.go
@@ -1,625 +1,0 @@
-package sqllog
-
-import (
-	"bytes"
-	"context"
-	"database/sql"
-	"errors"
-	"fmt"
-	"sync"
-	"time"
-
-	"github.com/canonical/k8s-dqlite/pkg/limited"
-	"github.com/sirupsen/logrus"
-	"go.etcd.io/etcd/api/v3/mvccpb"
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/metric"
-	"go.opentelemetry.io/otel/trace"
-)
-
-const (
-	otelName         = "sqllog"
-	SupersededCount  = 100
-	compactBatchSize = 1000
-	pollBatchSize    = 500
-)
-
-var (
-	otelTracer      trace.Tracer
-	otelMeter       metric.Meter
-	compactCnt      metric.Int64Counter
-	watcherGroupCnt metric.Int64Counter
-)
-
-func init() {
-	var err error
-	otelTracer = otel.Tracer(otelName)
-	otelMeter = otel.Meter(otelName)
-
-	compactCnt, err = otelMeter.Int64Counter(fmt.Sprintf("%s.compact", otelName), metric.WithDescription("Number of compact requests"))
-	if err != nil {
-		logrus.WithError(err).Warning("Otel failed to create create counter")
-	}
-	watcherGroupCnt, err = otelMeter.Int64Counter(fmt.Sprintf("%s.watcherGroup", otelName), metric.WithDescription("Number of watcherGroup requests"))
-	if err != nil {
-		logrus.WithError(err).Warning("Otel failed to create create counter")
-	}
-}
-
-type SQLLog struct {
-	mu sync.Mutex
-
-	config *SQLLogConfig
-
-	stop    func()
-	started bool
-
-	watcherGroups map[*watcherGroup]*watcherGroup
-	pollRevision  int64
-
-	notify chan int64
-	wg     sync.WaitGroup
-}
-
-type SQLLogConfig struct {
-	// Driver is the SQL driver to use to query the database.
-	Driver Driver
-
-	// CompactInterval is interval between database compactions performed by k8s-dqlite.
-	CompactInterval time.Duration
-
-	// PollInterval is the event poll interval used by k8s-dqlite.
-	PollInterval time.Duration
-
-	// WatchQueryTimeout is the timeout on the after query in the poll loop.
-	WatchQueryTimeout time.Duration
-}
-
-func New(config *SQLLogConfig) *SQLLog {
-	return &SQLLog{
-		config:        config,
-		notify:        make(chan int64, 100),
-		watcherGroups: make(map[*watcherGroup]*watcherGroup),
-	}
-}
-
-func (s *SQLLog) Start(startCtx context.Context) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	if s.started {
-		return nil
-	}
-
-	_, _, err := s.Create(startCtx, []byte("/registry/health"), []byte(`{"health":"true"}`), 0)
-	if err != nil {
-		return err
-	}
-
-	ctx, stop := context.WithCancel(context.Background())
-	if err := s.startWatch(ctx); err != nil {
-		stop()
-		return err
-	}
-
-	s.wg.Add(1)
-	go func() {
-		defer s.wg.Done()
-		s.ttl(ctx)
-	}()
-
-	s.stop = stop
-	s.started = true
-	return nil
-}
-
-func (s *SQLLog) Stop() error {
-	s.mu.Lock()
-
-	if !s.started {
-		return nil
-	}
-	s.stop()
-	s.stop, s.started = nil, false
-	s.mu.Unlock()
-
-	s.wg.Wait()
-	return nil
-}
-
-func (s *SQLLog) Close() error {
-	stopErr := s.Stop()
-	closeErr := s.config.Driver.Close()
-
-	return errors.Join(stopErr, closeErr)
-}
-
-func (s *SQLLog) compactStart(ctx context.Context) error {
-	currentRevision, err := s.config.Driver.CurrentRevision(ctx)
-
-	rows, err := s.config.Driver.AfterPrefix(ctx, []byte("compact_rev_key"), []byte("compact_rev_key\x00"), 0, currentRevision)
-	if err != nil {
-		return err
-	}
-
-	events, err := ScanAll(rows, scanEvent)
-	if err != nil {
-		return err
-	}
-
-	if len(events) == 0 {
-		_, _, err := s.Create(ctx, []byte("compact_rev_key"), nil, 0)
-		return err
-	} else if len(events) == 1 {
-		return nil
-	}
-
-	// this is to work around a bug in which we ended up with two compact_rev_key rows
-	maxRev := int64(0)
-	maxID := int64(0)
-	for _, event := range events {
-		if event.PrevKv != nil && event.PrevKv.ModRevision > maxRev {
-			maxRev = event.PrevKv.ModRevision
-			maxID = event.Kv.ModRevision
-		}
-	}
-
-	for _, event := range events {
-		if event.Kv.ModRevision == maxID {
-			continue
-		}
-		if err := s.config.Driver.DeleteRevision(ctx, event.Kv.ModRevision); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// DoCompact makes a single compaction run when called. It is intended to be called
-// from test functions that have access to the backend.
-func (s *SQLLog) DoCompact(ctx context.Context) (err error) {
-	ctx, span := otelTracer.Start(ctx, fmt.Sprintf("%s.DoCompact", otelName))
-	compactCnt.Add(ctx, 1)
-	defer func() {
-		span.RecordError(err)
-		span.End()
-	}()
-	if err := s.compactStart(ctx); err != nil {
-		return fmt.Errorf("failed to initialise compaction: %v", err)
-	}
-
-	// When executing compaction as a background operation
-	// it's best not to take too much time away from query
-	// operation and similar. As such, we do compaction in
-	// small batches. Given that this logic runs every second,
-	// on regime it should take usually just a couple batches
-	// to keep the pace.
-	start, target, err := s.config.Driver.GetCompactRevision(ctx)
-	if err != nil {
-		return err
-	}
-	span.SetAttributes(attribute.Int64("start", start))
-	// NOTE: Upstream is ignoring the last 1000 revisions, however that causes the following CNCF conformance test to fail.
-	// This is because of low activity, where the created list is part of the last 1000 revisions and is not compacted.
-	// Link to failing test: https://github.com/kubernetes/kubernetes/blob/f2cfbf44b1fb482671aedbfff820ae2af256a389/test/e2e/apimachinery/chunking.go#L144
-	// To address this, we only ignore the last 100 revisions instead
-	target -= SupersededCount
-	span.SetAttributes(attribute.Int64("target", target))
-	for start < target {
-		batchRevision := start + compactBatchSize
-		if batchRevision > target {
-			batchRevision = target
-		}
-		if err := s.config.Driver.Compact(ctx, batchRevision); err != nil {
-			return err
-		}
-		start = batchRevision
-	}
-	return nil
-}
-
-func (s *SQLLog) WatcherGroup(ctx context.Context) (limited.WatcherGroup, error) {
-	var err error
-
-	watcherGroupCnt.Add(ctx, 1)
-	ctx, span := otelTracer.Start(ctx, fmt.Sprintf("%s.WatcherGroup", otelName))
-	span.SetAttributes(attribute.Int64("currentRevision", s.pollRevision))
-	defer func() {
-		span.RecordError(err)
-		span.End()
-	}()
-
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	wg := &watcherGroup{
-		ctx:             ctx,
-		driver:          s.config.Driver,
-		currentRevision: s.pollRevision,
-		watchers:        make(map[int64]*watcher),
-		updates:         make(chan limited.WatcherGroupUpdate, 100),
-	}
-	stop := context.AfterFunc(ctx, func() {
-		s.mu.Lock()
-		defer s.mu.Unlock()
-		delete(s.watcherGroups, wg)
-		close(wg.updates)
-	})
-	wg.stop = stop
-	s.watcherGroups[wg] = wg
-	logrus.Debugf("created new WatcherGroup.")
-	return wg, nil
-}
-
-func (s *SQLLog) GetCompactRevision(ctx context.Context) (int64, int64, error) {
-	return s.config.Driver.GetCompactRevision(ctx)
-}
-
-func (s *SQLLog) List(ctx context.Context, key, rangeEnd []byte, limit, revision int64) (int64, []*limited.KeyValue, error) {
-	var err error
-
-	ctx, span := otelTracer.Start(ctx, fmt.Sprintf("%s.List", otelName))
-	defer func() {
-		span.RecordError(err)
-		span.End()
-	}()
-	span.SetAttributes(
-		attribute.String("key", string(key)),
-		attribute.String("rangeEnd", string(rangeEnd)),
-		attribute.Int64("limit", limit),
-		attribute.Int64("revision", revision),
-	)
-
-	compactRevision, currentRevision, err := s.config.Driver.GetCompactRevision(ctx)
-	if err != nil {
-		return 0, nil, err
-	}
-	if revision == 0 || revision > currentRevision {
-		revision = currentRevision
-	} else if revision < compactRevision {
-		return currentRevision, nil, limited.ErrCompacted
-	}
-
-	rows, err := s.config.Driver.List(ctx, key, rangeEnd, limit, revision)
-	if err != nil {
-		return 0, nil, err
-	}
-
-	result, err := ScanAll(rows, scanKeyValue)
-	if err != nil {
-		return 0, nil, err
-	}
-
-	return currentRevision, result, err
-}
-
-func (s *SQLLog) ttl(ctx context.Context) {
-	run := func(ctx context.Context, key []byte, revision int64, timeout time.Duration) {
-		select {
-		case <-ctx.Done():
-			return
-		case <-time.After(timeout):
-			s.Delete(ctx, key, revision)
-		}
-	}
-
-	s.wg.Add(1)
-	go func() {
-		defer s.wg.Done()
-
-		startRevision, err := s.config.Driver.CurrentRevision(ctx)
-		if err != nil {
-			logrus.Errorf("failed to read old events for ttl: %v", err)
-			return
-		}
-
-		rows, err := s.config.Driver.ListTTL(ctx, startRevision)
-		if err != nil {
-			logrus.Errorf("failed to read old events for ttl: %v", err)
-			return
-		}
-
-		var (
-			key             []byte
-			revision, lease int64
-		)
-		for rows.Next() {
-			if err := rows.Scan(&revision, &key, &lease); err != nil {
-				logrus.Errorf("failed to read old events for ttl: %v", err)
-				return
-			}
-			go run(ctx, key, revision, time.Duration(lease)*time.Second)
-		}
-
-		// TODO needs to be restarted, never cancelled/dropped
-		group, err := s.WatcherGroup(ctx)
-		if err != nil {
-			logrus.Errorf("failed to create watch group for ttl: %v", err)
-			return
-		}
-		group.Watch(1, []byte{0}, []byte{255}, startRevision)
-
-		for group := range group.Updates() {
-			for _, watcher := range group.Watchers() {
-				for _, event := range watcher.Events {
-					if event.Kv.Lease > 0 {
-						go run(ctx, []byte(event.Kv.Key), event.Kv.ModRevision, time.Duration(event.Kv.Lease)*time.Second)
-					}
-				}
-			}
-		}
-	}()
-}
-
-func (s *SQLLog) startWatch(ctx context.Context) error {
-	if err := s.compactStart(ctx); err != nil {
-		return err
-	}
-
-	pollInitialRevision, _, err := s.config.Driver.GetCompactRevision(ctx)
-	if err != nil {
-		return err
-	}
-	s.pollRevision = pollInitialRevision
-
-	// start compaction and polling at the same time to watch starts
-	// at the oldest revision, but compaction doesn't create gaps
-	s.wg.Add(2)
-
-	go func() {
-		defer s.wg.Done()
-
-		t := time.NewTicker(s.config.CompactInterval)
-
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			case <-t.C:
-				if err := s.DoCompact(ctx); err != nil {
-					logrus.WithError(err).Trace("compaction failed")
-				}
-			}
-		}
-	}()
-
-	go func() {
-		defer s.wg.Done()
-		s.poll(ctx)
-	}()
-
-	return nil
-}
-
-func (s *SQLLog) poll(ctx context.Context) {
-	ticker := time.NewTicker(s.config.PollInterval)
-	defer ticker.Stop()
-
-	waitForMore := true
-	for {
-		if waitForMore {
-			select {
-			case <-ctx.Done():
-				return
-			case check := <-s.notify:
-				if check <= s.pollRevision {
-					continue
-				}
-			case <-ticker.C:
-			}
-		}
-		events, err := s.getLatestEvents(ctx, s.pollRevision)
-		if err != nil {
-			if errors.Is(err, context.DeadlineExceeded) || errors.Is(err, context.Canceled) {
-				return
-			}
-			logrus.Errorf("fail to get latest events: %v", err)
-			continue
-		}
-
-		waitForMore = len(events) < 100
-		s.publishEvents(events)
-	}
-}
-
-func (s *SQLLog) publishEvents(events []*limited.Event) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if len(events) > 0 {
-		s.pollRevision = events[len(events)-1].Kv.ModRevision
-	}
-	for _, w := range s.watcherGroups {
-		if !w.publish(s.pollRevision, events) {
-			if w.stop() {
-				delete(s.watcherGroups, w)
-				close(w.updates)
-			}
-		}
-	}
-}
-
-func (s *SQLLog) getLatestEvents(ctx context.Context, pollRevision int64) ([]*limited.Event, error) {
-	var err error
-	watchCtx, cancel := context.WithTimeout(ctx, s.config.WatchQueryTimeout)
-	defer cancel()
-
-	watchCtx, span := otelTracer.Start(watchCtx, fmt.Sprintf("%s.getLatestEvents", otelName))
-	defer func() {
-		span.RecordError(err)
-		span.End()
-	}()
-	span.SetAttributes(attribute.Int64("pollRevision", pollRevision))
-
-	rows, err := s.config.Driver.After(watchCtx, pollRevision, pollBatchSize)
-	if err != nil {
-		return nil, err
-	}
-
-	events, err := ScanAll(rows, scanEvent)
-	if err != nil {
-		return nil, err
-	}
-	span.SetAttributes(attribute.Int("events", len(events)))
-	return events, nil
-}
-
-func (s *SQLLog) Count(ctx context.Context, key, rangeEnd []byte, revision int64) (int64, int64, error) {
-	var err error
-	ctx, span := otelTracer.Start(ctx, fmt.Sprintf("%s.Count", otelName))
-	defer func() {
-		span.RecordError(err)
-		span.End()
-	}()
-	span.SetAttributes(
-		attribute.String("key", string(key)),
-		attribute.String("rangeEnd", string(rangeEnd)),
-		attribute.Int64("revision", revision),
-	)
-
-	compactRevision, currentRevision, err := s.config.Driver.GetCompactRevision(ctx)
-	if err != nil {
-		return 0, 0, err
-	}
-	if revision == 0 || revision > currentRevision {
-		revision = currentRevision
-	} else if revision < compactRevision {
-		return currentRevision, 0, limited.ErrCompacted
-	}
-	count, err := s.config.Driver.Count(ctx, key, rangeEnd, revision)
-	if err != nil {
-		return 0, 0, err
-	}
-	return currentRevision, count, nil
-}
-
-func (s *SQLLog) Create(ctx context.Context, key, value []byte, lease int64) (int64, bool, error) {
-	rev, created, err := s.config.Driver.Create(ctx, key, value, lease)
-	if err != nil {
-		return 0, false, err
-	}
-	if created {
-		s.notifyWatcherPoll(rev)
-	}
-	return rev, created, nil
-}
-
-func (s *SQLLog) Delete(ctx context.Context, key []byte, revision int64) (rev int64, deleted bool, err error) {
-	rev, deleted, err = s.config.Driver.Delete(ctx, key, revision)
-	if err != nil {
-		return 0, false, err
-	}
-	if deleted {
-		s.notifyWatcherPoll(rev)
-	}
-	return rev, deleted, nil
-}
-
-func (s *SQLLog) Update(ctx context.Context, key []byte, value []byte, prevRev, lease int64) (rev int64, updated bool, err error) {
-	rev, updated, err = s.config.Driver.Update(ctx, key, value, prevRev, lease)
-	if err != nil {
-		return 0, false, err
-	}
-	if updated {
-		s.notifyWatcherPoll(rev)
-	}
-	return rev, updated, nil
-}
-
-func (s *SQLLog) notifyWatcherPoll(revision int64) {
-	select {
-	case s.notify <- revision:
-	default:
-	}
-}
-
-func (s *SQLLog) DbSize(ctx context.Context) (int64, error) {
-	var err error
-	ctx, span := otelTracer.Start(ctx, fmt.Sprintf("%s.DbSize", otelName))
-	defer func() {
-		span.RecordError(err)
-		span.End()
-	}()
-	size, err := s.config.Driver.GetSize(ctx)
-	span.SetAttributes(attribute.Int64("size", size))
-	return size, err
-}
-
-func ScanAll[T any](rows *sql.Rows, scanOne func(*sql.Rows) (T, error)) ([]T, error) {
-	var result []T
-	defer rows.Close()
-
-	for rows.Next() {
-		item, err := scanOne(rows)
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, item)
-	}
-
-	return result, nil
-}
-
-func scanKeyValue(rows *sql.Rows) (*limited.KeyValue, error) {
-	kv := &limited.KeyValue{}
-	err := rows.Scan(
-		&kv.ModRevision,
-		&kv.Key,
-		&kv.CreateRevision,
-		&kv.Lease,
-		&kv.Value,
-	)
-	if err != nil {
-		return nil, err
-	}
-	return kv, nil
-}
-
-func scanEvent(rows *sql.Rows) (*limited.Event, error) {
-	// Bundle in a single allocation
-	row := &struct {
-		event            limited.Event
-		kv, prevKv       limited.KeyValue
-		created, deleted bool
-	}{}
-	event := &row.event
-
-	err := rows.Scan(
-		&row.kv.ModRevision,
-		&row.kv.Key,
-		&row.created,
-		&row.deleted,
-		&row.kv.CreateRevision,
-		&row.prevKv.ModRevision,
-		&row.kv.Lease,
-		&row.kv.Value,
-		&row.prevKv.Value,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	event.Kv = &row.kv
-	if row.deleted {
-		row.event.Type = mvccpb.DELETE
-	}
-	if row.created {
-		event.Kv.CreateRevision = event.Kv.ModRevision
-	} else {
-		event.PrevKv = &row.prevKv
-		event.PrevKv.Key = event.Kv.Key
-		event.PrevKv.CreateRevision = event.Kv.CreateRevision
-		event.PrevKv.Lease = event.Kv.Lease
-	}
-	return event, nil
-}
-
-func appendEvents(src, dst []*limited.Event, key, rangeEnd []byte) []*limited.Event {
-	for _, event := range src {
-		if bytes.Compare(key, event.Kv.Key) <= 0 && bytes.Compare(rangeEnd, event.Kv.Key) > 0 {
-			dst = append(dst, event)
-		}
-	}
-	return dst
-}
